+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 18.07
 -----
 
+* A new API for editing fallbacks for a queue has been added:
+
+  * GET `/1.1/queues/<queue_id>/fallbacks`
+  * PUT `/1.1/queues/<queue_id>/fallbacks`
+
 * The entity of a user cannot be changed anymore. The following resource has been removed:
 
   * PUT `/1.1/users/<user_id>/entities/<entity_id>`